+++ app/models/long_text_answer.rb
@@ -1,0 +1,6 @@
+class LongTextAnswer < ActiveRecord::Base
+  self.implicit_order_column = "created_at"
+  belongs_to :step
+
+  validates :response, presence: true
+end