+++ app/models/short_text_answer.rb
@@ -1,6 +1,6 @@
 class ShortTextAnswer < ActiveRecord::Base
   self.implicit_order_column = "created_at"
-  belongs_to :question
+  belongs_to :step
 
   validates :response, presence: true
 end