+++ app/services/answer_factory.rb
@@ -1,15 +1,15 @@
 class AnswerFactory
-  attr_accessor :question
+  attr_accessor :step
 
-  def initialize(question:)
-    self.question = question
+  def initialize(step:)
+    self.step = step
   end
 
   def call
-    case question.contentful_type
+    case step.contentful_type
     when "radios" then RadioAnswer.new
     when "short_text" then ShortTextAnswer.new
-    else Answer.new
+    when "long_text" then LongTextAnswer.new
     end
   end
 end