+++ db/migrate/20201202150359_add_contentful_model_to_step.rb
@@ -1,0 +1,10 @@
+class AddContentfulModelToStep < ActiveRecord::Migration[6.0]
+  def up
+    add_column :steps, :contentful_model, :string
+    Step.update_all(contentful_model: "question")
+  end
+
+  def down
+    remove_column :steps, :contentful_model
+  end
+end