+++ spec/factories/answer.rb
@@ -1,19 +1,19 @@
 FactoryBot.define do
-  factory :answer do
-    association :question, factory: :question, contentful_type: "radios"
-
-    response { "Green" }
-  end
-
   factory :radio_answer do
-    association :question, factory: :question, contentful_type: "radios"
+    association :step, factory: :step, contentful_type: "radios"
 
     response { "Green" }
   end
 
   factory :short_text_answer do
-    association :question, factory: :question, contentful_type: "short_text"
+    association :step, factory: :step, contentful_type: "short_text"
 
     response { "Green" }
+  end
+
+  factory :long_text_answer do
+    association :step, factory: :step, contentful_type: "long_text"
+
+    response { "Lots of green" }
   end
 end