+++ spec/services/answer_factory_spec.rb
@@ -2,27 +2,27 @@ require "rails_helper"
 
 RSpec.describe AnswerFactory do
   describe "#call" do
-    context "when the question is for an unexpected type" do
-      it "returns a plain Answer object" do
-        question = create(:question, contentful_type: "foo")
-        result = described_class.new(question: question).call
-        expect(result).to be_kind_of(Answer)
-      end
-    end
-
-    context "when the question is for radios" do
+    context "when the step is for radios" do
       it "returns a new RadioAnswer object" do
-        question = create(:question, :radio)
-        result = described_class.new(question: question).call
+        step = create(:step, :radio)
+        result = described_class.new(step: step).call
         expect(result).to be_kind_of(RadioAnswer)
       end
     end
 
-    context "when the question is for short_text" do
+    context "when the step is for short_text" do
       it "returns a new ShortTextAnswer object" do
-        question = create(:question, :short_text)
-        result = described_class.new(question: question).call
+        step = create(:step, :short_text)
+        result = described_class.new(step: step).call
         expect(result).to be_kind_of(ShortTextAnswer)
+      end
+    end
+
+    context "when the step is for long_text" do
+      it "returns a new LongTextAnswer object" do
+        step = create(:step, :long_text)
+        result = described_class.new(step: step).call
+        expect(result).to be_kind_of(LongTextAnswer)
       end
     end
   end