+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [Unreleased]
 
+## [release-003] - 2020-12-07
+
+- add database foreign key constraints for better data integrity
+- Contentful entry requests use a new Redis read cache
+- service name included in the header
+- refactor name of "Plan" to "Journey"
+- refactor name of "Question" to "Step"
+- refactor redis caching into reusable class
+- users can be shown a step in the journey that is only static content
+
 ## [release-002] - 2020-11-16
 
 - migrate answer database table into Radio and ShortText
@@ -24,8 +34,10 @@ Contentful fixture
 - multiple radio questions can be answered in sequence
 - users can be asked to answer a short text question
 - Contentful can redirect users to preview endpoints
+- users can be asked to answer a long text question
 
-[unreleased]: https://github.com/DfE/DFE-Digital/buy-for-your-school/compare/release-001...HEAD
+[unreleased]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-003...HEAD
+[release-003]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-002...release-003
 [release-002]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-001...release-002
 [release-001]: https://github.com/DFE-Digital/buy-for-your-school/compare/release-000...release-001